cmptest: Move go test comparison into cmptest
This is a prefactor for allowing integration tests in
other packages, which is required for panic testing.

As part of this, we need to export the caller information in logs.
