+++ integration_test.go
@@ -1,26 +1,30 @@
-package faket
+package faket_test
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/prashantv/faket/internal/cmptest"
+)
 
 // These integration-style tests are used to compare the real output of running
 // a test to the result of `RunTest`.
 
 func TestCmp_Success(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("log1")
 		t.Log("log2")
 	})
 }
 
 func TestCmp_LogFormatting(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("a", 1, "b", 2, "c", "d")
 		t.Logf("a: %v b: %v", 1, 2)
 	})
 }
 
 func TestCmp_Skip(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("pre-skip")
 		t.Skip("skip")
 		t.Log("post-skip")
@@ -28,7 +32,7 @@ func TestCmp_Skip(t *testing.T) {
 }
 
 func TestCmp_Failure(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("pre-fail log")
 		t.Error("error log")
 		t.Log("post-fail log")
@@ -44,21 +48,21 @@ func TestCmp_Failure(t *testing.T) {
 // }
 
 func TestCmp_FailThenSkipCmp(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Error("error")
 		t.Skip("skipped")
 	})
 }
 
 func TestCmp_SkipThenFail(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Skip("skip")
 		t.Error("skipped error")
 	})
 }
 
 func TestCmp_Fatal(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("pre-fatal")
 		t.Fatal("fatal")
 		t.Log("post-fatal")
@@ -66,7 +70,7 @@ func TestCmp_Fatal(t *testing.T) {
 }
 
 func TestCmp_Cleanup(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("log 1")
 		t.Cleanup(func() {
 			t.Log("log in cleanup")
@@ -76,7 +80,7 @@ func TestCmp_Cleanup(t *testing.T) {
 }
 
 func TestCmp_CleanupError(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("log 1")
 		t.Cleanup(func() {
 			t.Error("error in cleanup")
@@ -86,7 +90,7 @@ func TestCmp_CleanupError(t *testing.T) {
 }
 
 func TestCmp_CleanupSkip(t *testing.T) {
-	compareTest(t, func(t testing.TB) {
+	cmptest.Compare(t, func(t testing.TB) {
 		t.Log("log 1")
 		t.Cleanup(func() {
 			t.Log("cleanup 1")
